BAVL-605: Split out the service layer into CourtBookingService and ProbationBookingService
