+++ server/routes/journeys/bookAVideoLink/court/handlers/bookingNotAvailableHandler.ts
@@ -5,8 +5,8 @@ import { IsNotEmpty, ValidateIf } from 'class-validator'
 import { parse } from 'date-fns'
 import { Page } from '../../../../../services/auditService'
 import { PageHandler } from '../../../../interfaces/pageHandler'
-import VideoLinkService from '../../../../../services/videoLinkService'
 import IsValidDate from '../../../../validators/isValidDate'
+import CourtBookingService from '../../../../../services/courtBookingService'
 
 const transformTime = (value: string) => (value ? parse(value, 'HH:mm', new Date(0)) : undefined)
 
@@ -57,13 +57,13 @@ export default class BookingNotAvailableHandler implements PageHandler {
 
   public BODY = Body
 
-  constructor(private readonly videoLinkService: VideoLinkService) {}
+  constructor(private readonly courtBookingService: CourtBookingService) {}
 
   public GET = async (req: Request, res: Response) => {
     const { user } = res.locals
     const { bookAVideoLink } = req.session.journey
 
-    const { availabilityOk, alternatives } = await this.videoLinkService.checkAvailability(bookAVideoLink, user)
+    const { availabilityOk, alternatives } = await this.courtBookingService.checkAvailability(bookAVideoLink, user)
 
     if (availabilityOk) {
       return res.redirect('check-booking')
