+++ server/routes/journeys/bookAVideoLink/court/handlers/newBookingHandler.ts
@@ -20,6 +20,7 @@ import PrisonService from '../../../../../services/prisonService'
 import PrisonerService from '../../../../../services/prisonerService'
 import VideoLinkService from '../../../../../services/videoLinkService'
 import { PrisonAppointment } from '../../../../../@types/bookAVideoLinkApi/types'
+import ReferenceDataService from '../../../../../services/referenceDataService'
 
 class Body {
   @Expose()
@@ -104,6 +105,7 @@ export default class NewBookingHandler implements PageHandler {
     private readonly courtsService: CourtsService,
     private readonly prisonService: PrisonService,
     private readonly prisonerService: PrisonerService,
+    private readonly referenceDataService: ReferenceDataService,
     private readonly videoLinkService: VideoLinkService,
   ) {}
 
@@ -121,7 +123,7 @@ export default class NewBookingHandler implements PageHandler {
 
     const rooms = await this.getRooms(prisoner.prisonId, bookingId, user)
 
-    const hearingTypes = await this.videoLinkService.getCourtHearingTypes(user)
+    const hearingTypes = await this.referenceDataService.getCourtHearingTypes(user)
 
     res.render('pages/bookAVideoLink/court/newBooking', {
       prisoner: {
