+++ server/services/index.ts
@@ -6,6 +6,9 @@ import ProbationTeamsService from './probationTeamsService'
 import PrisonService from './prisonService'
 import PrisonerService from './prisonerService'
 import VideoLinkService from './videoLinkService'
+import ReferenceDataService from './referenceDataService'
+import CourtBookingService from './courtBookingService'
+import ProbationBookingService from './probationBookingService'
 
 export const services = () => {
   const {
@@ -19,20 +22,26 @@ export const services = () => {
 
   const userService = new UserService(manageUsersApiClient, userPreferencesApiClient)
   const auditService = new AuditService(hmppsAuditClient)
+  const courtBookingService = new CourtBookingService(bookAVideoLinkApiClient)
   const courtsService = new CourtsService(bookAVideoLinkApiClient)
+  const probationBookingService = new ProbationBookingService(bookAVideoLinkApiClient)
   const probationTeamsService = new ProbationTeamsService(bookAVideoLinkApiClient)
   const prisonService = new PrisonService(bookAVideoLinkApiClient)
   const prisonerService = new PrisonerService(prisonerOffenderSearchApiClient)
   const videoLinkService = new VideoLinkService(bookAVideoLinkApiClient, prisonerOffenderSearchApiClient)
+  const referenceDataService = new ReferenceDataService(bookAVideoLinkApiClient)
 
   return {
     applicationInfo,
     userService,
     auditService,
+    courtBookingService,
     courtsService,
+    probationBookingService,
     probationTeamsService,
     prisonService,
     prisonerService,
+    referenceDataService,
     videoLinkService,
   }
 }
